Give a name!
Created a label system that still needs some working on but at least now
you can give your chip a name.  Future updates will include number
support and for chips... well hopefully chips that support looping,
first I will have to figure out if a chip has a loop in it and if it
does then find a way to create some sort of special loop table for it.