Big led change(No full chip yet)
So this is the beginning of leds, the ui is basically done, just a
glitch in the make button position.  Next is making the leds display on
chips, then after that is making a save system.  Probably after I make
them display on chips I will commit, after that should be a huge change
in the code to allow for saving boards